requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
openai==1.30.1
httpx>=0.25.0,<0.28
psycopg2-binary==2.9.9
pgvector==0.2.5
sqlalchemy==2.0.30
python-dotenv==1.0.1
tiktoken==0.7.0
slowapi==0.1.9
